Use decorations API to display gutter markers
This new API is designed exactly for our use case. One benefit is that
`GutterView.render` is now no longer called hundreds of times a second (!)

I copied most of the implementation from the git-diff package.

Test plan:

Opened a new editor and verified that gutter markers appear and disappear when
lint errors are caused and resolved. Also added a logging statement to
`GutterView.render` to verify that it's now only called when the file is
linted.