Consolidate markers into `LinterView`
Since `GutterView` and `HighlightsView` were basically the same thing, I
extracted out the shared code to `LinterView` and deleted them.

Test Plan:
Opened a file with lint errors. In another Atom window, went to settings and
turned gutter and highlights on and off and made sure that these settings still
updated in the other window.