Use the icon-right entity for gutter markers
Fixes #202.

Also is better because it doesn't make the gutter wider. Note that this
occupies the same spot as the fold marker, but that only shows up on hover
anyway, and it still appears.

Another change here is to begin using @text-color-warning and @text-color-error
which are provided by the theme.

Test Plan:
Looked at both error and warning level lint messages.