Stop Atom from crashing upon exit
In `deactivate`, `remove` is called, but before this change, `remove` doesn't
exist. The real fix for this is probably to make `InlineView` a real `View`,
but for now this will stop errors.

The fact that Atom crashes is a separate issue that I'll make a report for.

Test Plan:
Press `ctrl-cmd-opt l` and Atom doesn't crash