Make Linter load faster since all clients need to load this class.
Defers the loading of lodash and XRegExp until they are used.
Replaced _.memoize() with custom logic to pave the way to removing
the lodash dependency altogether in Linter.

The only other top-level `require()` calls that could be eliminated
are for `log` and `warn`.

One thing that I found a little odd is that, assuming `@executablePath`
never changes over the lifetime of the `Linter` once it is set,
`@_statCache` will only ever have one entry in it.