Filter messages to only display those related to the current file
In certain systems, for instance the package manager cargo for
the rust language, an entire package with multiple files must
be built at once and all the errors are reported in bulk.

Previously, it was impossible to filter these messages based on the
file they apply to and only show the ones relevant to the current file.
This commit provides the hooks for linters to say which file a message
belongs to (though this is optional, to maintain backwards compatibility)
so that we can filter out all messages not related to the current file.

By default, all messages are assumed to be relevant to the current file.